+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ## Unreleased
 
 * Update Twitter logo ([PR #4011](https://github.com/alphagov/govuk_publishing_components/pull/4011))
+* Use Terser instead of Uglifier ([PR #3990](https://github.com/alphagov/govuk_publishing_components/pull/3990))
 
 ## 38.2.0