+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 
 ## Unreleased
 
+## 38.3.1
+
 * Remove Oxford comma from metadata component ([PR #4012](https://github.com/alphagov/govuk_publishing_components/pull/4012))
 * Add null check for input element ([PR #4022](https://github.com/alphagov/govuk_publishing_components/pull/4022))